+++ packages/orq-rc/src/orq_ai_sdk/models/createconversationop.py
@@ -199,7 +199,7 @@ class CreateConversationResponseBody(BaseModel):
     r"""Unix timestamp in milliseconds when the conversation was last modified."""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "conv_01kfg3pf4w34190t60zvjshvft"
+        "conv_01kfgavgepcdrac4gerrprp1ds"
     )
     r"""Unique conversation identifier with `conv_` prefix."""