+++ packages/orq-rc/src/orq_ai_sdk/models/createdatasourceop.py
@@ -314,7 +314,7 @@ class CreateDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KFG3PFMHVYGZ0X091KP9SF80"
+        "01KFGAVH0X9JQZ02VWPHMXTX1J"
     )
     r"""The unique identifier of the data source"""