+++ packages/orq-rc/src/orq_ai_sdk/models/retrievedatasourceop.py
@@ -90,7 +90,7 @@ class RetrieveDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KFG3PFMGMV2257D6CKKKCG6A"
+        "01KFGAVH0VPGWN6ADCBVGDYX7S"
     )
     r"""The unique identifier of the data source"""