+++ packages/orq-rc/src/orq_ai_sdk/models/retrieveidentityop.py
@@ -103,7 +103,7 @@ class RetrieveIdentityResponseBody(BaseModel):
     created: Optional[datetime] = None
     r"""The date and time the resource was created"""
 
-    updated: Optional[datetime] = parse_datetime("2026-01-21T11:05:24.540Z")
+    updated: Optional[datetime] = parse_datetime("2026-01-21T13:10:29.486Z")
     r"""The date and time the resource was last updated"""
 
     @model_serializer(mode="wrap")