+++ packages/orq-rc/RELEASES.md
@@ -4929,3 +4929,13 @@ Based on:
 - [python v4.2.0-rc.55] packages/orq-rc
 ### Releases
 - [PyPI v4.2.0-rc.55] https://pypi.org/project/orq-ai-sdk/4.2.0-rc.55 - packages/orq-rc
+
+## 2026-01-21 13:13:12
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.691.0 (2.797.1) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [python v4.2.0-rc.56] packages/orq-rc
+### Releases
+- [PyPI v4.2.0-rc.56] https://pypi.org/project/orq-ai-sdk/4.2.0-rc.56 - packages/orq-rc